Break up LibraryPage into smaller, self-contained pieces
LibraryPage now has become a few other pieces:

**SidePanel**

Contains logic around mobile and desktop display of the EmbeddedSidePanel (the filters)

**ResumableContentGrid**

Shows, conditionally, the resumable content for the user. Tests were basically moved over to the new component 1-to-1.

**SearchResultsGrid**

Much like ResumableContentGrid, this was migrated over with most test 1-to-1.

**useCopies** was created to allow shared store and simple API for managing the displayedCopies

**BUGFIX** - Embedded side panel didn't disable the buttons it ought to have, this is fixed in that file's only changes
