+++ kolibri/plugins/learn/assets/src/composables/__tests__/useCopies.spec.js
@@ -1,0 +1,22 @@
+import useCopies from '../useCopies';
+
+describe('useCopies', () => {
+  it('should default to an empty array', () => {
+    const { displayedCopies } = useCopies();
+    expect(displayedCopies).toEqual([]);
+  });
+  it('should set displayed copies to value passed to setCopies function', () => {
+    const { displayedCopies, setCopies } = useCopies();
+    setCopies([{ id: '1' }, { id: '2' }]);
+    expect(displayedCopies.value).toEqual([{ id: '1' }, { id: '2' }]);
+  });
+  it('displayedCopies should be a module-level store, so the previous test should still pass', () => {
+    const { displayedCopies } = useCopies();
+    expect(displayedCopies.value).toEqual([{ id: '1' }, { id: '2' }]);
+  });
+  it('should reset displayedCopies to an empty array when clearCopies is called', () => {
+    const { displayedCopies, clearCopies } = useCopies();
+    clearCopies();
+    expect(displayedCopies.value).toEqual([]);
+  });
+});
