+++ kolibri/plugins/learn/assets/src/views/LibraryPage/index.vue
@@ -1,0 +1,250 @@
+<template>
+
+  <div>
+    <main
+      class="main-grid"
+      :style="gridOffset"
+    >
+      <div v-if="!windowIsLarge">
+        <KButton
+          icon="filter"
+          data-test="filter-button"
+          :text="translator.$tr('filter')"
+          :primary="false"
+          @click="toggleSidePanelVisibility"
+        />
+      </div>
+      <!--
+        - If search is loading, show loader.
+        - If there are no search results, show channels and resumable
+        content.
+        - Otherwise, show search results.
+      -->
+      <KCircularLoader
+        v-if="searchLoading"
+        class="loader"
+        type="indeterminate"
+        :delay="false"
+      />
+      <div v-else-if="!displayingSearchResults">
+        <h2>{{ coreString('channelsLabel') }}</h2>
+        <ChannelCardGroupGrid
+          v-if="rootNodes.length"
+          data-test="channel-cards"
+          class="grid"
+          :contents="rootNodes"
+        />
+        <!-- ResumableContentGrid handles whether it renders or not internally -->
+        <ResumableContentGrid />
+      </div>
+
+      <!-- SearchResultsGrid handles whether it renders or not internally -->
+      <SearchResultsGrid :currentCardViewStyle="currentCardViewStyle" />
+    </main>
+
+    <!-- Side Panels for filtering and searching  -->
+    <SidePanel
+      :mobileSidePanelIsOpen="mobileSidePanelIsOpen"
+      @toggleMobileSidePanel="toggleSidePanelVisibility"
+    />
+  </div>
+
+</template>
+
+
+<script>
+
+  import { mapState } from 'vuex';
+
+  import responsiveWindowMixin from 'kolibri.coreVue.mixins.responsiveWindowMixin';
+  import commonCoreStrings from 'kolibri.coreVue.mixins.commonCoreStrings';
+  import FilterTextbox from 'kolibri.coreVue.components.FilterTextbox';
+  import { crossComponentTranslator } from 'kolibri.utils.i18n';
+  import useSearch from '../../composables/useSearch';
+  import useLearnerResources from '../../composables/useLearnerResources';
+  import commonLearnStrings from '../commonLearnStrings';
+  import ChannelCardGroupGrid from '../ChannelCardGroupGrid';
+  import ResumableContentGrid from './ResumableContentGrid';
+  import SearchResultsGrid from './SearchResultsGrid';
+  import SidePanel from './SidePanel';
+
+  export default {
+    name: 'LibraryPage',
+    metaInfo() {
+      return {
+        title: this.learnString('learnLabel'),
+      };
+    },
+    components: {
+      ChannelCardGroupGrid,
+      ResumableContentGrid,
+      SearchResultsGrid,
+      SidePanel,
+    },
+    mixins: [commonLearnStrings, commonCoreStrings, responsiveWindowMixin],
+    setup() {
+      const {
+        searchTerms,
+        displayingSearchResults,
+        searchLoading,
+        moreLoading,
+        results,
+        more,
+        labels,
+        search,
+        searchMore,
+        removeFilterTag,
+        clearSearch,
+        setCategory,
+      } = useSearch();
+      const {
+        resumableContentNodes,
+        moreResumableContentNodes,
+        fetchMoreResumableContentNodes,
+      } = useLearnerResources();
+      return {
+        searchTerms,
+        displayingSearchResults,
+        searchLoading,
+        moreLoading,
+        results,
+        more,
+        labels,
+        search,
+        searchMore,
+        removeFilterTag,
+        clearSearch,
+        setCategory,
+        resumableContentNodes,
+        moreResumableContentNodes,
+        fetchMoreResumableContentNodes,
+      };
+    },
+    data: function() {
+      return {
+        currentCardViewStyle: 'card',
+        mobileSidePanelIsOpen: false,
+      };
+    },
+    computed: {
+      ...mapState(['rootNodes']),
+      sidePanelWidth() {
+        if (this.windowIsSmall || this.windowIsMedium) {
+          return 0;
+        } else if (this.windowBreakpoint < 4) {
+          return 234;
+        } else {
+          return 346;
+        }
+      },
+      gridOffset() {
+        return this.isRtl
+          ? { marginRight: `${this.sidePanelWidth + 24}px` }
+          : { marginLeft: `${this.sidePanelWidth + 24}px` };
+      },
+    },
+    watch: {
+      searchTerms() {
+        this.mobileSidePanelIsOpen = false;
+      },
+      mobileSidePanelIsOpen() {
+        if (this.mobileSidePanelIsOpen) {
+          document.documentElement.style.position = 'fixed';
+          return;
+        }
+        document.documentElement.style.position = '';
+      },
+    },
+    created() {
+      this.search();
+      this.translator = crossComponentTranslator(FilterTextbox);
+    },
+    methods: {
+      toggleSidePanelVisibility() {
+        this.mobileSidePanelIsOpen = !this.mobileSidePanelIsOpen;
+      },
+    },
+    $trs: {
+      /* eslint-disable kolibri/vue-no-unused-translations */
+      results: {
+        message: '{results, number, integer} {results, plural, one {result} other {results}}',
+        context: 'Number of results for a given term after a Library search.',
+      },
+      moreThanXResults: {
+        message: 'More than {results} results',
+        context: 'Number of results for a given term after a Library search.',
+      },
+      /* eslint-disable kolibri/vue-no-unused-translations */
+      viewAsList: {
+        message: 'View as list',
+        context: 'Label for a button used to view resources as a list.',
+      },
+      viewAsGrid: {
+        message: 'View as grid',
+        context: 'Label for a button used to view resources as a grid.',
+      },
+      clearAll: {
+        message: 'Clear all',
+        context: 'Clickable link which removes all currently applied search filters.',
+      },
+    },
+  };
+
+</script>
+
+
+<style lang="scss" scoped>
+
+  .main-grid {
+    margin-top: 40px;
+    margin-right: 24px;
+  }
+
+  $gutters: 24px;
+
+  .card-grid-item {
+    margin-bottom: $gutters;
+  }
+
+  .loader {
+    margin-top: 60px;
+  }
+
+  .toggle-view-buttons {
+    float: right;
+  }
+
+  .full-screen-side-panel {
+    position: relative;
+    width: 100vw;
+  }
+
+  .results-title {
+    display: inline-block;
+    margin-bottom: 24px;
+  }
+
+  .results-header-group {
+    margin-bottom: 24px;
+  }
+
+  .filter-action-button {
+    display: inline-block;
+    margin: 4px;
+    margin-left: 8px;
+  }
+
+  .side-panel-chips {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: flex-start;
+    margin-bottom: -8px;
+    margin-left: -8px;
+  }
+
+  .chip {
+    margin-bottom: 8px;
+    margin-left: 8px;
+  }
+
+</style>
